requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1

# Vector database
chromadb>=0.4.0

# Data processing
biopython>=1.81
numpy>=1.24.0

# Quantization
bitsandbytes>=0.41.0

# Utilities
tqdm>=4.66.0

